+++ app_fenrir/src/main/kotlin/dev/ragnarok/fenrir/activity/ChatActivityBubbles.kt
@@ -90,7 +90,7 @@ class ChatActivityBubbles : NoMainActivity(), PlaceProvider, AppStyleable {
             Place.CHAT -> {
                 val peer: Peer = args.getParcelableCompat(Extra.PEER) ?: return
                 val chatFragment =
-                    newInstance(args.getInt(Extra.ACCOUNT_ID), args.getInt(Extra.OWNER_ID), peer)
+                    newInstance(args.getLong(Extra.ACCOUNT_ID), args.getLong(Extra.OWNER_ID), peer)
                 attachToFront(chatFragment)
             }
             Place.VK_PHOTO_ALBUM_GALLERY, Place.FAVE_PHOTOS_GALLERY, Place.SIMPLE_PHOTO_GALLERY, Place.VK_PHOTO_TMP_SOURCE, Place.VK_PHOTO_ALBUM_GALLERY_SAVED, Place.VK_PHOTO_ALBUM_GALLERY_NATIVE -> newInstance(
@@ -118,7 +118,7 @@ class ChatActivityBubbles : NoMainActivity(), PlaceProvider, AppStyleable {
             Place.DOC_PREVIEW -> {
                 val document: Document? = args.getParcelableCompat(Extra.DOC)
                 if (document != null && document.hasValidGifVideoLink()) {
-                    val aid = args.getInt(Extra.ACCOUNT_ID)
+                    val aid = args.getLong(Extra.ACCOUNT_ID)
                     val documents = ArrayList(listOf(document))
                     val extra = GifPagerActivity.buildArgs(aid, documents, 0)
                     place.launchActivityForResult(this, GifPagerActivity.newInstance(this, extra))
@@ -156,8 +156,8 @@ class ChatActivityBubbles : NoMainActivity(), PlaceProvider, AppStyleable {
         val data = intent
         if (data != null && data.extras != null) {
             NotificationHelper.setBubbleOpened(
-                (data.extras ?: return).getInt(Extra.ACCOUNT_ID),
-                (data.extras ?: return).getInt(Extra.OWNER_ID)
+                (data.extras ?: return).getLong(Extra.ACCOUNT_ID),
+                (data.extras ?: return).getLong(Extra.OWNER_ID)
             )
         }
         super.onResume()
@@ -210,7 +210,7 @@ class ChatActivityBubbles : NoMainActivity(), PlaceProvider, AppStyleable {
         const val ACTION_OPEN_PLACE = "dev.ragnarok.fenrir.activity.ChatActivityBubbles.openPlace"
 
 
-        fun forStart(context: Context?, accountId: Int, peer: Peer): Intent {
+        fun forStart(context: Context?, accountId: Long, peer: Peer): Intent {
             val intent = Intent(context, ChatActivityBubbles::class.java)
             intent.action = ACTION_OPEN_PLACE
             intent.putExtra(Extra.PLACE, PlaceFactory.getChatPlace(accountId, accountId, peer))
