+++ app_fenrir/src/main/kotlin/dev/ragnarok/fenrir/activity/LocalJsonToChatActivity.kt
@@ -109,7 +109,7 @@ class LocalJsonToChatActivity : NoMainActivity(), PlaceProvider, AppStyleable {
             Place.DOC_PREVIEW -> {
                 val document: Document? = args.getParcelableCompat(Extra.DOC)
                 if (document != null && document.hasValidGifVideoLink()) {
-                    val aid = args.getInt(Extra.ACCOUNT_ID)
+                    val aid = args.getLong(Extra.ACCOUNT_ID)
                     val documents = ArrayList(listOf(document))
                     val extra = GifPagerActivity.buildArgs(aid, documents, 0)
                     place.launchActivityForResult(this, GifPagerActivity.newInstance(this, extra))
